[MLOP-225] KeyFeature need dtype to be a required arg (#126)

+++ tests/unit/butterfree/core/transform/conftest.py
@@ -189,7 +189,7 @@ def feature3(spark_context, spark_session):
 
 @fixture
 def key_id():
-    return KeyFeature(name="id", description="description")
+    return KeyFeature(name="id", description="description", dtype=DataType.INTEGER)
 
 
 @fixture
